Refactor FTXUI runloop to handle refreshing synced realm data
