+++ sync-todo/v2/client/cpp/screens/error-modal.hpp
@@ -6,7 +6,7 @@
 #include "ftxui/component/screen_interactive.hpp"  // for Component, ScreenInteractive
 #include "ftxui/dom/elements.hpp"  // for text, hbox, separator, Element, operator|, vbox, border
 
-#include "../display-screen.hpp"
+#include "display-screen.hpp"
 
 class ErrorModal {
 private:
