--- sync-todo/v2/client/cpp/screens/static-dashboard.hpp
@@ -1,4 +1,0 @@
-class StaticDashboard {
- public:
-  void init();
-};
